refactor: split combo/chat.ts per 800-line new-file cap
- Extracted auto-strategy logic to chat/auto.ts
- Extracted execution loop to chat/handler.ts
- Extracted target executor to chat/executor.ts
- Extracted routing/init to chat/routing.ts
- Shared types in chat/types.ts
- chat.ts now barrel re-export + handleComboChat orchestrator

All files <800 lines. Fixed import paths for nested module.
