+++ open-sse/services/combo/chat/auto.ts
@@ -21,28 +21,28 @@ import {
   getIntentConfig,
   mapIntentToTaskType,
   filterTargetsByRequestCompatibility,
-} from "../context.ts";
-import { isRecord } from "../utils.ts";
+} from "../../context.ts";
+import { isRecord } from "../../utils.ts";
 import {
   resolveResetWindowConfig,
   resolveSlaRoutingPolicy,
   orderTargetsByResetAwareQuota,
   orderTargetsByResetWindow,
   preScreenTargets,
-} from "../quota.ts";
-import { _registerExecutionCandidates } from "../state.ts";
-import { dedupeTargetsByExecutionKey } from "../dag.ts";
+} from "../../quota.ts";
+import { _registerExecutionCandidates } from "../../state.ts";
+import { dedupeTargetsByExecutionKey } from "../../dag.ts";
 import {
   scheduleShadowRouting,
   resolveShadowTargets,
-} from "../shadow.ts";
+} from "../../shadow.ts";
 import { orderTargetsByEvalScores } from "../../evalRouting.ts";
 import { generateRoutingHints } from "../../manifestAdapter";
 import {
   expandAutoComboCandidatePool,
   buildAutoCandidates,
   scoreAutoTargets,
-} from "../auto.ts";
+} from "../../auto.ts";
 import { parseModel } from "../../model.ts";
 import {
   getModelContextLimitForModelString,
@@ -50,14 +50,14 @@ import {
   sortTargetsByUsage,
   sortTargetsByCost,
   sortTargetsByContextSize,
-} from "../sorting.ts";
+} from "../../sorting.ts";
 import {
   fisherYatesShuffle,
   getNextFromDeck,
 } from "../../../../src/shared/utils/shuffleDeck";
 import {
   comboModelNotFoundResponse,
-} from "../constants.ts";
+} from "../../constants.ts";
 import type {
   ResolvedComboTarget,
   PreScreenResult,
