+++ open-sse/services/combo/chat/routing.ts
@@ -10,14 +10,14 @@ import {
   resolveComboConfig,
   getDefaultComboConfig,
   resolveComboTargetTimeoutMs,
-} from "../comboConfig.ts";
+} from "../../comboConfig.ts";
 import {
   resolveContextRelayConfig,
   resolveUniversalHandoffConfig,
   SKIP_UNIVERSAL_HANDOFF_FLAG,
-} from "../contextHandoff.ts";
+} from "../../contextHandoff.ts";
 import { getLastSessionModel } from "../../../../src/lib/db/contextHandoffs.ts";
-import { applyComboAgentMiddleware } from "../comboAgentMiddleware.ts";
+import { applyComboAgentMiddleware } from "../../comboAgentMiddleware.ts";
 import { normalizeRoutingStrategy } from "../../../../src/shared/constants/routingStrategies.ts";
 import {
   resolveResilienceSettings,
